requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
aiosqlite==0.19.0

# 爬虫
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
lxml==4.9.3

# AI服务
openai>=1.0.0

# 定时任务
apscheduler==3.10.4

# 工具库
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-docx==1.1.0
